Split lib.rs file into many smaller files
