+++ src/core/capture/activated/iterator.rs
@@ -1,7 +1,10 @@
-use crate::Activated;
-use crate::Capture;
-use crate::Error;
-use crate::PacketCodec;
+use crate::{
+    core::{
+        capture::{Activated, Capture},
+        codec::PacketCodec,
+    },
+    Error,
+};
 
 /// Implement an Iterator of Packet
 pub struct PacketIter<S: Activated + ?Sized, C> {
