+++ src/core/codec.rs
@@ -1,4 +1,4 @@
-use crate::Packet;
+use crate::core::packet::Packet;
 
 /// This trait is used to implement Stream and Iterator feature.
 /// This is almost like `map()`.
