+++ src/sendqueue/mod.rs
@@ -1,0 +1,4 @@
+#[cfg(windows)]
+pub mod windows;
+#[cfg(windows)]
+pub use windows as sendqueue;
